+++ fresh.gen.ts
@@ -21,9 +21,8 @@ import * as $15 from './routes/signin.tsx';
 import * as $16 from './routes/signout.tsx';
 import * as $$0 from './islands/Entry/index.tsx';
 import * as $$1 from './islands/EntryInput/index.tsx';
-import * as $$2 from './islands/EntryList/index.tsx';
-import * as $$3 from './islands/EntryQuery/index.tsx';
-import * as $$4 from './islands/TagUpdater/index.tsx';
+import * as $$2 from './islands/EntryQuery/index.tsx';
+import * as $$3 from './islands/TagUpdater/index.tsx';
 
 const manifest = {
   routes: {
@@ -48,9 +47,8 @@ const manifest = {
   islands: {
     './islands/Entry/index.tsx': $$0,
     './islands/EntryInput/index.tsx': $$1,
-    './islands/EntryList/index.tsx': $$2,
-    './islands/EntryQuery/index.tsx': $$3,
-    './islands/TagUpdater/index.tsx': $$4,
+    './islands/EntryQuery/index.tsx': $$2,
+    './islands/TagUpdater/index.tsx': $$3,
   },
   baseUrl: import.meta.url,
 };